* CC: fix a endless loop crash bug when parsing C99 designated initializer
Forum discussion is here:
http://forums.codeblocks.org/index.php/topic,21417.0.html

When parsing
struct AAA a1 = {.x = 1, .y=2}; // first line
struct AAA a2 = {.x = 1, .y=2}; // second line
When we see a "dot", we need to skip to "}", thus the close brace is
skipped, and the open brace is just start a function call of DoParse().
If we goes to the second line, we will call another DoParse(), so there
is recursive calls, and finally goes to crash.

Also, a cc test case is added