* CCTest: automatically test for doxygen document. (thanks White-Tiger)
This is SF ticket 232:

Here is the note from White-Tiger
1) I've also fixed some trimming while I was at it.
(wxString::Trim(bool trim_right) defaults to true and cctest did two
Trim().. one with default true and one with implicit true)
2) the code inside the constructor isn't required since it's usually
the default anyway... but maybe do it implicit nontheless
3) the patch to src/plugins/codecompletion/parser/parser_base.cpp
should be properly redone, or at least every use of it verified
since it'll probably always default to storeDocumentation = true even
in situations where it "shouldn't"
4) the last "test" should fail... there are 2 special test values for
doxygen comments, "*" as a "non-empty" wildcard, and "-" to check for
empty docs

Also, some fixes from my(olldbg) side
1, enable the store doxygen doc option for the parser in cctest.
2, fix a bug in the the cctest's parser, we should firstly mark the file
in the TokenTree, so that the doxygen document will bestored in the
Token::m_Doc, otherwise, it was wrongly stored in the m_ImpDoc. See
Token* ParserThread::DoAddToken() about how the doxygen document are added
to the Token.

Discussion link: https://sourceforge.net/p/codeblocks/tickets/232/